pyproject.toml
[project]
name = "hive"
version = "0.1.0"
description = "Shared persistent memory MCP server for Claude agents and teams"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.78",
    "fastapi>=0.135.2",
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "mangum>=0.21.0",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "boto3-stubs>=1.42.78",
    "moto>=5.1.22",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["jose.*", "moto.*"]
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hive"]
